Refactored into smaller functions and added individual training dirs.
The first portion of this commit refactors the if __main__ section at
the bottom of the script into smaller functions. This makes it easier to
read.

The second portion changes how logs checkpoints and generations are
saved. They are now saved into their own training run directory, the
name of which can be specified with the --training_run argument. This
makes it easier to do multiple training runs because there will be no
save conflicts as long as the name of each run is different.