Modified heightmap dataset volume mount in docker-compose.yaml.
This is not ideal because other people will have their datasets in
different locations and will have to modify the mount path. One solution
would be to expect datasets to be in the local ./data/ directory, but
this also isn't ideal since they take up quite a bit of space and it
would be annoying to duplicate them if they're used in multiple
projects. I also tried symlinking a global datasets directory into
./data/datasets/ but docker didn't mount it. I think symlinking would be
the best solution here if I the mounting issue can be solved. If not
then some sort of configuration script may be required.