Added version of terra with less agressive time conditioning.
The time embedding is now resized and added as a channel at the start of
the network. This is how it was done in previous NTG. I'm making this
in hopes of achieving faster convergence.

This version also uses adamw and has b2 set to 0.95 in order to prevent
loss spikes (someone on twitter recommended this).