Fixed issue #1.
The issue was caused by the initial state being saved in the partial
sampling function. When training from scratch this meant that the state
was saved in its randomly initialized form, and when loaded from a
checkpoint the state was saved as that checkpoint. This explains why
training from scratch only produced pure noise samples while training
from a checkpoint did not.